transfer: widen View-size multiplications to size_t
clang-tidy bugprone-implicit-widening-of-multiplication-result (treated
as error) flagged int*int products used to size 64-bit Views in the
intersection/MC RHS integrators and the field-exchange-planner test.
Perform the multiplication in size_t via an explicit cast.

+++ src/pcms/transfer/omega_h_intersection_rhs_integrator.cpp
@@ -108,9 +108,10 @@ OmegaHIntersectionRHSIntegrator::BuildDataImpl(
   // Pass 2: fill coords, node_gids, and coeffs on device. node_gids/coeffs hold
   // ndof (target DOFs per element) entries per integration point.
   Kokkos::View<Real**, DeviceMemorySpace> coords("rhs_coords", num_pts, 2);
-  Kokkos::View<PetscInt*, DeviceMemorySpace> node_gids("rhs_node_gids",
-                                                       num_pts * ndof);
-  Kokkos::View<Real*, DeviceMemorySpace> coeffs("rhs_coeffs", num_pts * ndof);
+  Kokkos::View<PetscInt*, DeviceMemorySpace> node_gids(
+    "rhs_node_gids", static_cast<std::size_t>(num_pts) * ndof);
+  Kokkos::View<Real*, DeviceMemorySpace> coeffs(
+    "rhs_coeffs", static_cast<std::size_t>(num_pts) * ndof);
 
   Kokkos::parallel_for(
     "rhs_fill", nelems, KOKKOS_LAMBDA(int elm) {
